requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
gunicorn
bcrypt
werkzeug
PyJWT
